Lots of small items: - All images (except one) removed from project - Image details are now stored in json in ./js/data/backgrounds.js - Images will try to load remotely:   - if they fail, they'll fall back to the built-in image (see config.js)   - if the built-in image fails, the screen still looks nice (uses gradient)   - screen desperately needs fade in transition to avoid FOUC (https://en.wikipedia.org/wiki/Flash_of_unstyled_content) - Properties that read newTabState moved into stats component - Code retrieving about links moved into footer component - Components not using state were updated to inherit from ImmutableComponent - Clock updated to start with a time (no longer waits 1 second before showing)
Auditors: @cezaraugusto, @bbondy

When auditing, please notice that the ONLY way I could find to detect image load
failure was to put a hidden image tag with an onerror handler. Having a div with
a `background-image` CSS attribute doesn't appear to offer a way to catch the error.
You can provide a fallback image URL, but the title of the image won't match
(and down the road, the author details may not match, if there are multiple authors)